Disable bypassing via CLI options

+++ DATA/tools.py
@@ -197,6 +197,8 @@ async def check_bypass(
 
     returns bool: true means continue with operation
     """
+    if not ctx.bot.bypassing:
+        return False
     if ctx.author.id in ctx.bot.auto_bypass:
         await asyncio.sleep(1) # Guilded is weird here, since it will occasionally return 404 Message Not Found
         await msg.delete()
